refactor: 경로 설정 로직을 커스텀 훅으로 분리
